+++ src/index.ts
@@ -1,3 +1,4 @@
 export * from './create-simulation';
 export * from './types';
 export * from './test-helpers';
+export * from './test-helpers-with-explicit-react';
